@@ -1,39 +1,55 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
+import { AfterViewInit, Component, OnInit, ViewChild } from '@angular/core';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
-import { IEnvelopedData, IHostOnboarding, IHostOnboardingProcess } from '@eventi/interfaces';
+import { IEnvelopedData, IHostOnboarding } from '@eventi/interfaces';
 import { AdminService } from "src/app/services/admin.service";
 import { ICacheable } from 'src/app/app.interfaces';
+import { BaseAppService } from 'src/app/services/app.service';
 
 @Component({
   selector: 'app-admin-onboarding-list',
   templateUrl: './admin-onboarding-list.component.html',
   styleUrls: ['./admin-onboarding-list.component.scss']
 })
-export class AdminOnboardingListComponent implements OnInit {
-  
-  public onboardingRequests;
-  public displayedColumns: string[] = ['state', 'last_modified', 'host', 'onboarding_page'];
-  public onboardingTableData:ICacheable<IEnvelopedData<IHostOnboarding[], void>> = {
+export class AdminOnboardingListComponent implements OnInit, AfterViewInit {
+  @ViewChild(MatPaginator) paginator: MatPaginator;
+
+  onboardingsDataSrc:MatTableDataSource<IHostOnboarding>;
+  displayedColumns: string[] = ['host', 'state', 'last_modified', 'onboarding_page'];
+  onboardings:ICacheable<IEnvelopedData<IHostOnboarding[], void>> = {
     data: null,
     loading: false,
     error: ""
-}
-  @ViewChild(MatPaginator) paginator: MatPaginator;
+  }
 
-  constructor(private adminService: AdminService) { }
+  constructor(private adminService: AdminService, private appService:BaseAppService) { }
 
-  async ngOnInit() {
-    await this.getOnboardingProcesses();  
+  get pager():MatPaginator { return this.onboardingsDataSrc.paginator }
+
+  ngOnInit() {
+    this.onboardingsDataSrc = new MatTableDataSource<IHostOnboarding>([]);
+    this.getOnboardingProcesses();
   }
 
   ngAfterViewInit() {
-    this.onboardingRequests.paginator = this.paginator;
+    this.onboardingsDataSrc.paginator = this.paginator;
   }
 
   async getOnboardingProcesses() {
-    this.onboardingTableData.data = await this.adminService.readOnboardingProcesses();
-    this.onboardingRequests = new MatTableDataSource<IHostOnboarding>(this.onboardingTableData.data.data);
+    this.onboardings.loading = true;
+    return this.adminService.readOnboardingProcesses(this.pager?.pageIndex, this.pager?.pageSize)
+      .then(d => {
+        this.onboardings.data = d;
+        this.onboardingsDataSrc.data = d.data;
+        if(this.pager) {
+          this.pager.length = d.__paging_data.total;
+        }
+      })
+      .catch(e => this.onboardings.error = e)
+      .finally(() => this.onboardings.loading = false)
   }
 
+  openOnboarding(onboarding:IHostOnboarding) {
+    this.appService.navigateTo(`/admin/onboarding/${onboarding._id}`);
+  }
 }
